Moved AccessType into its on file and a new 'core' subdirectory.
